sprinkles::config
Struct Scoop

pub struct Scoop {

    pub use_external_7zip: bool,
    pub use_lessmsi: bool,
    pub no_junction: bool,
    pub scoop_repo: ScoopRepo,
    pub scoop_branch: ScoopBranch,
    pub proxy: Option<Proxy>,
    pub autostash_on_conflict: bool,
    pub default_architecture: Architecture,
    pub debug: bool,
    pub force_update: bool,
    pub show_update_log: bool,
    pub show_manifest: bool,
    pub shim: ScoopShim,
    pub root_path: PathBuf,
    pub global_path: PathBuf,
    pub cache_path: Option<PathBuf>,
    pub gh_token: Option<String>,
    pub virustotal_api_key: Option<String>,
    pub ignore_running_processes: bool,
    pub hold_update_until: Option<String>,
    pub use_isolated_path: IsolatedPath,
    pub last_update: Option<String>,
    pub aria2_config: Config,
    /* private fields */

}

Scoop configuration

Fields
use_external_7zip: bool

External 7zip (from path) will be used for archives extraction

use_lessmsi: bool

Prefer lessmsi utility over native msiexec

no_junction: bool

The ‘current’ version alias will not be used. Shims and shortcuts will point to specific version instead

scoop_repo: ScoopRepo

Git repository containing the scoop adaptor’s source code

This configuration is useful for custom forks of scoop, or a scoop replacement

scoop_branch: ScoopBranch

Allow to use different branch than master

Could be used for testing specific functionalities before released into all users

If you want to receive updates earlier to test new functionalities use develop (see: https://github.com/ScoopInstaller/Scoop/issues/2939)

proxy: Option<Proxy>

By default, we will use the proxy settings from Internet Options, but with anonymous authentication.

To use the credentials for the current logged-in user, use ‘currentuser’ in place of username:password

To use the system proxy settings configured in Internet Options, use ‘default’ in place of host:port

An empty or unset value for proxy is equivalent to ‘default’ (with no username or password)

To bypass the system proxy and connect directly, use ‘none’ (with no username or password)

autostash_on_conflict: bool

When a conflict is detected during updating, Scoop will auto-stash the uncommitted changes. (Default is false, which will abort the update)

default_architecture: Architecture

Allow to configure preferred architecture for application installation

If not specified, architecture is determined by system

debug: bool

Additional and detailed output will be shown

force_update: bool

Force apps updating to bucket’s version

show_update_log: bool

Show update log

show_manifest: bool

Displays the manifest of every app that’s about to be installed, then asks user if they wish to proceed

shim: ScoopShim

Choose scoop shim build

root_path: PathBuf

Path to Scoop root directory

global_path: PathBuf

Path to Scoop root directory for global apps

cache_path: Option<PathBuf>

For downloads, defaults to ‘cache’ folder under Scoop root directory

gh_token: Option<String>

GitHub API token used to make authenticated requests

This is essential for checkver and similar functions to run without incurring rate limits and download from private repositories

virustotal_api_key: Option<String>

API key used for uploading/scanning files using virustotal

See: https://support.virustotal.com/hc/en-us/articles/115002088769-Please-give-me-an-API-key

ignore_running_processes: bool

When set to false (default), Scoop would stop its procedure immediately if it detects any target app process is running. Procedure here refers to reset/uninstall/update.

When set to true, Scoop only displays a warning message and continues procedure.

hold_update_until: Option<String>

Disable/Hold Scoop self-updates, until the specified date scoop hold scoop will set the value to one day later

Should be in the format ‘YYYY-MM-DD’, ‘YYYY/MM/DD’ or any other forms that accepted by ‘[System.DateTime]::Parse()’

Ref: https://docs.microsoft.com/dotnet/api/system.datetime.parse?view=netframework-4.5

use_isolated_path: IsolatedPath

When set to true (default), Scoop will use SCOOP_PATH environment variable to store apps’ PATHs.

When set to arbitrary non-empty string, Scoop will use that string as the environment variable name instead. This is useful when you want to isolate Scoop from the system PATH.

last_update: Option<String>

The timestamp of the last scoop update

aria2_config: Config

Aria2 configuration

Implementations

impl Scoop

pub fn load() -> Result<Self>

Converts the config path into the Scoop struct

Errors

The file was not valid UTF-8

The read file was did not match the expected structure

pub fn get_path() -> PathBuf

Gets the scoop config path

Panics

The config directory does not exist

pub fn update_last_update_time(&mut self)

Update the last time the scoop was updated

pub fn save(&self) -> Result<()>

Save the modified scoop config

Errors

The struct could not be serialized to JSON

The file could not be written

pub fn make_strict(&mut self)

Make the config strict

This will remove all fields that are not in the config struct

pub fn to_object(&self) -> Result<Value>

Convert the config to a JSON object

Errors

The config could not be converted to a JSON object

pub fn from_object(object: Value) -> Result<Self>

Convert the JSON object to Scoop config

Errors

The JSON object could not be deserialized to a Scoop config