sprinkles::contexts
Trait ScoopContext

pub trait ScoopContext:
    Clone
    + Send
    + Sync
    + 'static {
    type Config;

    const APP_NAME: &'static str;
    const ELEVATED: bool;
    const CONTEXT_NAME: &'static str = Self::APP_NAME;

    // Required methods
    fn config(&self) -> &Self::Config;
    fn config_mut(&mut self) -> &mut Self::Config;
    fn symlinks_enabled(&self) -> bool;
    fn proxy(&self) -> Option<&Proxy>;
    fn path(&self) -> &Path;
    fn logging_dir(&self) -> Result<PathBuf>;
    fn open_repo(&self) -> Option<Result<Repo>>;
    fn outdated(&self) -> impl Future<Output = Result<bool, Error>>;

    // Provided methods
    fn sub_path(&self, segment: impl AsRef<Path>) -> PathBuf { ... }
    fn apps_path(&self) -> PathBuf { ... }
    fn buckets_path(&self) -> PathBuf { ... }
    fn cache_path(&self) -> PathBuf { ... }
    fn persist_path(&self) -> PathBuf { ... }
    fn shims_path(&self) -> PathBuf { ... }
    fn workspace_path(&self) -> PathBuf { ... }
    fn scripts_path(&self) -> PathBuf { ... }
    fn installed_apps(&self) -> Result<Vec<PathBuf>> { ... }
    fn app_installed(&self, name: impl AsRef<str>) -> Result<bool> { ... }
    fn context_app_path(&self) -> PathBuf { ... }
    fn known_buckets(&self) -> &Map<&'static str, &'static str> { ... }

}

An adapter for Scoop-like contexts

This is used to provide a common interface for Scoop-like contexts, and to allow for mocking in tests

Generally, you should not call this type directly, but instead use the User or Global types, or another type that implements ScoopContext.

Example

let context = User::new().unwrap();
let scoop_path = context.path();

Required Associated Constants

const APP_NAME: &'static str

The name of the context

This is used internally to ignore this app when searching for installed apps, and to display the context name in the output.

This string should match what the app’s name is in Scoop, if applicable.

const ELEVATED: bool

Whether the context requires elevation

This should be true if the context requires elevation to run (i.e a global context) and false if it does not (i.e a user context).

Provided Associated Constants

const CONTEXT_NAME: &'static str = Self::APP_NAME

The name of the context

This should be the actual name of the context, not the name of the app. For example, the name of the global context should be “global”, and the name of the user context should be “user”.

This will default to the value of ScoopContext::APP_NAME, but can be overridden.

Required Associated Types

type Config

The config used by the given context

Required Methods

fn config(&self) -> &Self::Config

Get a reference to the context’s configuration

fn config_mut(&mut self) -> &mut Self::Config

Get a mutable reference to the context’s configuration

fn symlinks_enabled(&self) -> bool

Check if symlinks are enabled

Generally this is an option in the config, and should be passed directly from the config.

fn proxy(&self) -> Option<&Proxy>

Get the proxy for the context

Generally this is an option in the config, and should be passed directly from the config.

fn path(&self) -> &Path

Gets the context’s path

fn logging_dir(&self) -> Result<PathBuf>

Get the path to the log directory

fn open_repo(&self) -> Option<Result<Repo>>

Open the context’s app repository, if any

fn outdated(&self) -> impl Future<Output = Result<bool, Error>>

Check if the context is outdated

This may manifest in different ways, depending on the context

For User and Global, this will check if the Scoop repository is outdated and return true if it is.

For implementors, this should return true if your provider is outdated. For example, the binary which hooks into this trait will return true if there is a newer version available.

Provided Methods

fn sub_path(&self, segment: impl AsRef<Path>) -> PathBuf

Get a sub path within the context’s path

This function will attempt to create the path if it does not exist but will not panic or return an error if it fails

fn apps_path(&self) -> PathBuf

Get the contexts’s apps path

fn buckets_path(&self) -> PathBuf

Get the contexts’s buckets path

fn cache_path(&self) -> PathBuf

Get the contexts’s cache path

fn persist_path(&self) -> PathBuf

Get the contexts’s persist path

fn shims_path(&self) -> PathBuf

Get the contexts’s shims path

fn workspace_path(&self) -> PathBuf

Get the contexts’s workspace path

fn scripts_path(&self) -> PathBuf

Get the contexts’s scripts path

fn installed_apps(&self) -> Result<Vec<PathBuf>>

List all scoop apps and return their paths, except for the context’s app

fn app_installed(&self, name: impl AsRef<str>) -> Result<bool>

Checks if the app is installed by its name

fn context_app_path(&self) -> PathBuf

Get the path to the context’s app

This should return the path to the app’s directory, not the repository.

For example, if the context is the user context, this should return the path to the scoop app

fn known_buckets(&self) -> &Map<&'static str, &'static str>

List all known buckets

This function will return a map of bucket name to url

Errors

Reading the buckets file

Deserializing the buckets file

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.