sprinkles::git::clone

Module progress

Modules

bytesize

ByteSize is a semantic wrapper for byte count representations.

prodash

Prodash is a dashboard for displaying the progress of concurrent application.

Structs

BoxedDynNestedProgress

An opaque type for storing DynNestedProgress.

Discard

An implementation of NestedProgress which discards all calls.

DoOrDiscard

An implementation of Progress which can be created easily from Option<impl Progress>.

DynNestedProgressToNestedProgress

A bridge type that implements NestedProgress for any type that implements DynNestedProgress.

Read

A structure passing every read call through to the contained Progress instance using inc_by(bytes_read).

Task

The value associated with a spot in the hierarchy.

ThroughputOnDrop

Emit a message with throughput information when the instance is dropped.

Unit

A configurable and flexible unit for use in Progress::init().

Value

Progress associated with some item in the progress tree.

Write

A structure passing every write call through to the contained Progress instance using inc_by(bytes_written).

Enums

Either

An implementation of NestedProgress showing either one or the other implementation.

MessageLevel

The severity of a message

Constants

UNKNOWN

The default Id to use if there is no need for an id.

Traits

Count

A thread-safe read-only counter, with unknown limits.

DynNestedProgress

An object-safe trait for describing hierarchical progress.

NestedProgress

A trait for describing hierarchical progress.

Progress

A trait for describing non-hierarchical progress.

Functions

bytes

A unit for displaying bytes with throughput and progress percentage.

count

A unit for displaying human readable numbers with throughput and progress percentage, and a single decimal place.

count_with_decimals

A unit for displaying human readable numbers with name suffix, with throughput and progress percentage, and decimals decimal places.

steps

A predefined unit for displaying a multi-step progress

Type Aliases

AtomicStep

The amount of steps a progress can make, for threadsafe counting.

Id

Four bytes of function-local unique and stable identifier for each item added as progress, like b“TREE“ or b“FILE“.

Step

The amount of steps a progress can make

StepShared

As step, but shareable.