sprinkles::git::clone::progress::prodash::progress

Module key

Structs

Adjacency

A type providing information about what’s above and below Tree items.

Key

A type identifying a spot in the hierarchy of Tree items.

Enums

SiblingLocation

Determines if a sibling is above or below in the given level of hierarchy

Type Aliases

Level

a level in the hierarchy of key components