sprinkles::handles::packages
Struct PackageHandle

pub struct PackageHandle<'a, C> { /* private fields */ }

A package handle

Implementations

impl<'a, C: ScoopContext> PackageHandle<'a, C>

pub async fn new(ctx: &'a C, reference: Reference) -> Result<Self>

Create a new package handle

Errors

The package reference is invalid

The package is not installed

The package was not installed correctly

The package’s remote manifest could not be found or parsed

The package’s install directory could not be found

pub async fn from_manifest(ctx: &'a C, manifest: &Manifest) -> Result<Self>

Create a new package handle from a manifest

Errors

The manifest was not provided a bucket

The manifest was not provided a name

Any further errors from PackageHandle::new

pub fn current(&self) -> PathBuf

Get the package’s install path

This will return either the current folder, or the version directory, if the current folder is not linked or the config has no_junction set to true

pub fn remote_manifest(&self) -> &Manifest

Get the package’s remote manifest

pub fn local_manifest(&self) -> Result<Manifest>

Get the package’s manifest

Errors

Loading and parsing the manifest failed

pub fn install_manifest(&self) -> Result<InstallManifest>

Get the package’s install manifest

Errors

Loading and parsing the install manifest failed

pub fn persist_dir(&self) -> PathBuf

Get the package’s persist directory

pub fn version_dir(&self) -> PathBuf

Get the package’s current version directory

This will return the version of the remote manifest, or the version of the package reference, if specified

pub fn local_version(&self) -> Result<Version>

Get the package’s local version

Errors

See more at PackageHandle::local_manifest

pub fn list_versions(&self) -> Result<Vec<VersionHandle>, Error>

List all versions of the package

Errors

Reading the package’s versions failed

pub fn unlink_current(&self) -> Result<()>

Unlink the current folder

Errors

Unlinking the current folder failed

pub fn link_current(&self) -> Result<()>

Link the current folder to the package’s install folder

This will do nothing if the config has no_junction set to true

Errors

Unlinking the current folder failed

Linking the current folder failed

The current folder is not a symlink

pub fn version_paths(&self) -> Result<Rc<[PathBuf]>>

Get the package’s version paths

Errors

Reading the package’s version paths failed

pub unsafe fn name(&self) -> &str

Get the package’s name

Safety

This field is manually set in the remote manifest, and by default is uninitialized. This may cause undefined behavior.

Use Manifest::name_opt or, to ensure that this function returns properly, use the CreateManifest trait to set the name, or create the manifest, rather than other methods that might fail to set the name.

pub fn reference(&self) -> &Reference

Get the package’s reference

pub fn running(&self) -> bool

Check if the package handle owns a running process