sprinkles::hash
Struct Hash

pub struct Hash { /* private fields */ }

Hash result type

Implementations

impl Hash

pub fn no_prefix(&self) -> &str

Get the hash with no prefix

pub fn hash_type(&self) -> HashType

Get the hash type

impl Hash

pub fn from_hex(bytes: &[u8]) -> Self

Create a new hash from hash hex bytes

pub async fn get_for_app( ctx: &impl ScoopContext, manifest: &Manifest, arch: Architecture, ) -> Result<Vec<Hash>, Error>

Get a hash for an app

Note that this function expects the current url to be the updated version. In the case that we have to download and compute the hash, we will download the app from the current url, not the autoupdate url.

Errors

If the hash is not found

If the hash is invalid

If the hash mode is invalid

If the URL is invalid

If the source is invalid

If the JSON is invalid

If the hash is not found in the headers

If the hash is not found in the RDF

If the hash is not found in the XML

If the hash is not found in the text

If the hash is not found in the JSON

pub fn compute(reader: impl BufRead, hash_type: HashType) -> Hash

Compute a hash from a source

pub fn from_rdf( source: impl AsRef<[u8]>, file_name: impl AsRef<str>, ) -> Result<Hash, Error>

Parse a hash from an RDF source

Errors

If the hash is not found

pub fn find_hash_in_xml( source: impl AsRef<str>, substitutions: &SubstitutionMap, xpath: impl AsRef<str>, ) -> Result<Hash, Error>

Parse a hash from an XML source

Errors

If the hash is not found

If the hash is invalid

If the XML is invalid

If the XPath is invalid

pub fn from_text( source: impl AsRef<str>, substitutions: &SubstitutionMap, regex: impl AsRef<str>, ) -> Result<Hash, Error>

Parse a hash from a text source

Errors

If the hash is not found

If the hash is invalid

pub fn from_json( source: impl AsRef<[u8]>, substitutions: &SubstitutionMap, json_path: impl AsRef<str>, ) -> Result<Hash, Error>

Parse a hash from a json source

Errors

If the hash is not found

If the hash is invalid

pub fn find_hash_in_headers( _headers: &HeaderMap<HeaderValue>, ) -> Result<Hash, Error>

Find a hash in the headers of a response

Errors

peepeepoopoo