sprinkles::packages::models::manifest
Struct Manifest

pub struct Manifest {

    pub empty: Option<StringArray>,
    pub schema: Option<String>,
    pub comment: Option<StringArray>,
    pub architecture: Option<ManifestArchitecture>,
    pub autoupdate: Option<Autoupdate>,
    pub cookie: Option<HashMap<String, Option<Value>>>,
    pub depends: Option<SingleOrArray<Reference>>,
    pub description: Option<String>,
    pub extract_to: Option<StringArray>,
    pub homepage: Option<String>,
    pub innosetup: Option<bool>,
    pub license: Option<PackageLicense>,
    pub notes: Option<StringArray>,
    pub persist: Option<NestedArray<String>>,
    pub psmodule: Option<Psmodule>,
    pub suggest: Option<Suggest>,
    pub version: Version,
    pub env_add_path: Option<StringArray>,
    pub env_set: Option<HashMap<String, Option<Value>>>,
    pub install_config: InstallConfig,
    /* private fields */

}

The manifest for a package

Fields
empty: Option<StringArray>

A comment.

schema: Option<String>

The schema of the manifest

comment: Option<StringArray>
👎Deprecated: Use ## instead

A comment.

architecture: Option<ManifestArchitecture>

The architecture of the package

autoupdate: Option<Autoupdate>

The autoupdate configuration

cookie: Option<HashMap<String, Option<Value>>>

Undocumented: Found at https://github.com/se35710/scoop-java/search?l=JSON&q=cookie

depends: Option<SingleOrArray<Reference>>

The dependencies of the package

description: Option<String>

The description of the package

extract_to: Option<StringArray>

Extract to dir or dirs

homepage: Option<String>

The homepage of the package

innosetup: Option<bool>

True if the installer InnoSetup based. Found in https://github.com/ScoopInstaller/Main/search?l=JSON&q=innosetup

license: Option<PackageLicense>

The license of the package

notes: Option<StringArray>

The manifest notes

persist: Option<NestedArray<String>>

Directories to persist when updating

psmodule: Option<Psmodule>

The PowerShell module of the package

suggest: Option<Suggest>

The suggested dependencies of the package

version: Version

The version of the package

env_add_path: Option<StringArray>

The environment variables to add to PATH

env_set: Option<HashMap<String, Option<Value>>>

The environment variables to set

install_config: InstallConfig

The install configuration

Implementations

impl Manifest

pub unsafe fn name(&self) -> &str

Get the name of the manifest

Safety

This field is manually set, and by default is uninitialized. This may cause undefined behavior.

Use Manifest::name_opt or, to ensure that this function returns properly, use the [CreateManifest] trait to set the name, or create the manifest, rather than other methods that might fail to set the name.

pub fn name_opt(&self) -> Option<&str>

Get the name of the manifest, or None if it is not set

pub fn set_name(&mut self, name: impl Into<String>)

Set the name of the manifest

pub unsafe fn bucket(&self) -> &str

Get the bucket the manifest is from

Safety

This field is manually set, and by default is uninitialized. This may cause undefined behavior.

Use Manifest::bucket_opt or, to ensure that this function returns properly, use the [CreateManifest] trait to set the bucket, or create the manifest, rather than other methods that might fail to set the bucket.

pub fn bucket_opt(&self) -> Option<&str>

Get the bucket the manifest is from, or None if it is not set

pub fn set_bucket(&mut self, bucket: impl Into<String>)

Set the bucket the manifest is from

impl Manifest

pub fn install_config(&self, arch: Architecture) -> InstallConfig

Get the install config for a given architecture

pub fn autoupdate_config(&self, arch: Architecture) -> Option<AutoupdateConfig>

Get the autoupdate config for the default architecture

pub fn download_urls(&self, arch: Architecture) -> Option<Vec<DownloadUrl>>

Get the download urls for a given architecture

pub fn with_bucket(self, bucket: &Bucket) -> Self

Apply a bucket to a manifest

pub fn depends(&self) -> Vec<Reference>

List the dependencies of a given manifest, in the order that they will be installed

Note that this does not include the package itself as a dependency

pub fn from_reference( ctx: &impl ScoopContext, (bucket, name): (String, String), ) -> Result<Self>

Gets the manifest from a bucket and manifest name

Errors

If the manifest doesn’t exist or bucket is invalid

pub fn binary_matches( &self, regex: &Regex, arch: Architecture, ) -> Option<Vec<String>>

Check if the manifest binaries matche the given regex

pub fn list_installed(ctx: &impl ScoopContext) -> Result<Vec<Result<Self>>>

List all installed app manifests

Errors

Invalid install manifest

Reading directories fails

Panics

If the file name is invalid

pub fn is_installed( &self, ctx: &impl ScoopContext, bucket: Option<&str>, ) -> bool

Check if the manifest is installed

pub async fn set_version( &mut self, ctx: &impl ScoopContext, version: String, ) -> Result<(), Error>

Set the manifest version and get the hash for the manifest

Errors

Missing autoupdate field

Hash error

pub fn commit_message_matches(&self, commit: &Commit<'_>) -> bool

Check if the commit’s message matches the name of the manifest

pub fn commit_diff_matches(&self, commit: &Commit<'_>) -> Result<bool>

Check if the commit’s changed files matches the name of the manifest

Errors

Git2 errors

pub fn last_updated_info( &self, ctx: &impl ScoopContext, ) -> Result<(Option<DateTime<FixedOffset>>, Option<Signature>)>

Get the time and author of the commit where this manifest was last changed

Errors

Invalid bucket

Invalid repo bucket

Internal git2 errors

pub fn install_manifest( &self, ctx: &impl ScoopContext, ) -> Result<InstallManifest>

Get InstallManifest for Manifest

Errors

Missing or invalid InstallManifest