sprinkles::packages

Trait MergeDefaults

pub trait MergeDefaults {
    type Default;

    // Required method
    fn merge_default(
        &self,
        default: Self::Default,
        arch: Architecture,
    ) -> Self::Default;
}

Merge defaults for a given architecture and the provided field

Required Associated Types

type Default

Output & Input type

Required Methods

fn merge_default( &self, default: Self::Default, arch: Architecture, ) -> Self::Default

Merge the architecture specific autoupdate config with the arch agnostic one