sprinkles

Module scripts

Powershell Script helpers

This module provides a way to create and run Powershell scripts

Example

let script = PowershellScript::new("Write-Host 'Hello, world!'");
let runner = script.save_to(ctx.scripts_path()).unwrap();
runner.run().unwrap();

Modules

installer

Utilities for running installers and uninstallers

summary

Summaries of what will run for a hook or installer

Structs

PowershellScript

A powershell script

ScriptRunner

A script runner

Enums

Error

Errors that can occur when running a script

Type Aliases

Result

A Powershell script runner result